+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed 🔧
 
+* Center window titles.
 * Tweak size and alignment of some emojis to match other text.
 
 ### Fixed 🐛